APITEST/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djongo==1.3.6
gunicorn==20.1.0
Markdown==3.3.4
pymongo==3.12.1
pytz==2021.3
sqlparse==0.2.4
